Made make_data not to be imported on each call of knmt
