+++ nmt_chainer/dataprocessing/make_data.py
@@ -14,6 +14,7 @@ import json
 import operator
 import os.path
 import gzip
+import sys
 
 from nmt_chainer.utilities.utils import ensure_path
 # import h5py
@@ -666,50 +667,6 @@ def build_dataset_with_align_info(src_fn, tgt_fn, align_fn,
                                                        )
 
 
-def define_parser(parser):
-    parser.add_argument(
-        "src_fn", help="source language text file for training data")
-    parser.add_argument(
-        "tgt_fn", help="target language text file for training data")
-    parser.add_argument(
-        "save_prefix", help="created files will be saved with this prefix")
-    parser.add_argument("--align_fn", help="align file for training data")
-    parser.add_argument("--src_voc_size", type=int, default=32000,
-                        help="limit source vocabulary size to the n most frequent words")
-    parser.add_argument("--tgt_voc_size", type=int, default=32000,
-                        help="limit target vocabulary size to the n most frequent words")
-#     parser.add_argument("--add_to_valid_set_every", type = int)
-#     parser.add_argument("--shuffle", default = False, action = "store_true")
-#     parser.add_argument("--enable_fast_shuffle", default = False, action = "store_true")
-
-    parser.add_argument("--max_nb_ex", type = int, help = "only use the first MAX_NB_EX examples")
-    
-    parser.add_argument("--test_src", help = "specify a source test set")
-    parser.add_argument("--test_tgt", help = "specify a target test set")
-    
-    parser.add_argument("--dev_src", help = "specify a source dev set")
-    parser.add_argument("--dev_tgt", help = "specify a target dev set")
-    parser.add_argument("--mode_align", choices = ["unk_align", "all_align"])
-    parser.add_argument("--use_voc", help = "specify an exisiting vocabulary file")
-    
-    parser.add_argument("--tgt_segmentation_type", choices = ["word", "word2char", "char"], default = "word")
-    parser.add_argument("--src_segmentation_type", choices = ["word", "word2char", "char"], default = "word")
-    
-    
-    
-def cmdline(arguments=None):
-    import sys
-    import argparse
-    parser = argparse.ArgumentParser(description="Prepare training data.",
-                                     formatter_class=argparse.ArgumentDefaultsHelpFormatter)
-    
-    define_parser(parser)
-    
-    args = parser.parse_args(args = arguments)
-    
-    do_make_data(args)
-    
-    
 def do_make_data(args):
     
     
@@ -718,7 +675,7 @@ def do_make_data(args):
         sys.exit(1)
 
     if not ((args.dev_src is None) == (args.dev_tgt is None)):
-        print >>sys.stderr, "Command Line Error: either specify both --test_src and --test_tgt or neither"
+        print >>sys.stderr, "Command Line Error: either specify both --dev_src and --dev_tgt or neither"
         sys.exit(1)
 
     save_prefix_dir, save_prefix_fn = os.path.split(args.save_prefix)
